[#2131] Throw warning message when design filename contains illegal characters
