+++ core/resources/l10n/messages.properties
@@ -1341,6 +1341,10 @@ SaveRktWarningDialog.donotshow=Do not show this dialog again
 saveAs.openrocket.title=Save as OpenRocket ork file
 saveAs.rocksim.title=Export as RockSim rkt file
 
+! SaveAsFileChooser
+SaveAsFileChooser.illegalFilename.title = Illegal filename
+SaveAsFileChooser.illegalFilename.message = The filename '%s' may not contain the character ' %c '. Please remove it.
+
 ! StorageOptionChooser
 StorageOptChooser.lbl.Simdatatostore = Simulated data to store:
 StorageOptChooser.rdbut.Allsimdata = All simulated data
