+++ swing/src/net/sf/openrocket/gui/main/BasicFrame.java
@@ -86,6 +86,7 @@ import net.sf.openrocket.gui.util.OpenFileWorker;
 import net.sf.openrocket.gui.util.SaveFileWorker;
 import net.sf.openrocket.gui.util.SwingPreferences;
 import net.sf.openrocket.gui.util.URLUtil;
+import net.sf.openrocket.gui.widgets.SaveFileChooser;
 import net.sf.openrocket.l10n.Translator;
 import net.sf.openrocket.logging.Markers;
 import net.sf.openrocket.rocketcomponent.ComponentChangeEvent;
@@ -1410,7 +1411,7 @@ public class BasicFrame extends JFrame {
 	public boolean exportRockSimAction() {
 		File file;
 
-		final SaveAsFileChooser chooser = SaveAsFileChooser.build(document, FileType.ROCKSIM);
+		final DesignFileSaveAsFileChooser chooser = DesignFileSaveAsFileChooser.build(document, FileType.ROCKSIM);
 
 		int option = chooser.showSaveDialog(BasicFrame.this);
 
@@ -1507,9 +1508,14 @@ public class BasicFrame extends JFrame {
 	private boolean saveAsAction() {
 		File file = null;
 
-		final SaveAsFileChooser chooser = SaveAsFileChooser.build(document, FileType.OPENROCKET);
+		final DesignFileSaveAsFileChooser chooser = DesignFileSaveAsFileChooser.build(document, FileType.OPENROCKET);
 
 		int option = chooser.showSaveDialog(BasicFrame.this);
+
+		// If the user entered an illegal filename, show the dialog again
+		while (option == SaveFileChooser.ILLEGAL_FILENAME_ERROR) {
+			option = chooser.showSaveDialog(BasicFrame.this);
+		}
 
 		if (option != JFileChooser.APPROVE_OPTION) {
 			log.info(Markers.USER_MARKER, "User decided not to save, option=" + option);
